Fixed the full screen dying issue when full screen is invoked with no stream selected.
Chat ListView is VERY raw.  This is on purpose so the Chat folks can mold to their plan.

The Login tab is essentaily a shell.  This is also on purpose to the Login group can mold to their plan.

The Stream Select screen is an independent Activity.  This seems way heavy on the surface, BUT that choice was made to support using a grid view or map select option in the future.  If a view flipper was used, then the main app would be playng with the map feed the whole time, which is a waste.

The actual Alerts are NOT implemented.  We will simply send JSON alert message to the NotificationManager.  The NotificationManager can send notices to the system bar, as Dialogs, or post to the phone's opening screen (like ESPN and email alerts do).  There is some thought needed here since System bar notifications can't be seen in full screen mode, alert dialogs are needed.